feat: add additional password to create and delete products
As the admin username and password are exposed so anyone can see the pages only the admin has
access, I decided to add a very simple password authentication to prevent anyone from adding or
deleting products. Currently, it is not a secure method, just a thin layer of security. I may
increase the security in the future.

+++ pages/api/products/[id].js
@@ -1,6 +1,7 @@
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
+const password = process.env.PRODUCTS_PASSWORD;
 
 export default async function productHandler(req, res) {
   const {
@@ -18,14 +19,19 @@ export default async function productHandler(req, res) {
     if (!product) {
       res.status(404).json({ error: "Product not found" });
       return;
-    }
+    };
 
     res.status(200).json(product);
   } else if (method === 'DELETE') {
+    if (req.headers.authorization !== password) {
+      res.status(401).json({ error: "Unauthorized" });
+      return;
+    };
+
     if (!product) {
       res.status(404).json({ error: "Product not found" });
       return;
-    }
+    };
 
     await prisma.product.delete({
       where: { id: productId },
